Rename the compat test suite for clarity
vncservers.py -> utils.py: it's shared infrastructure, not compat-suite
code -- every functional test file imports it (cli, proxy, replay,
roundtrip, scene_player, events, api_lifecycle, pixel_format), not just
the two server-compat files. The old name implied it belonged to the
compat suite specifically.

test_servers.py / test_os_servers.py -> test_server_compat_docker.py /
test_server_compat_native.py: names the axis explicitly (containerized
vs raw-on-the-host) instead of leaving it to be inferred from which
fleet each file imports, and drops the 'os' framing in favour of what
actually distinguishes them.

The two stay separate files rather than merging (their only code
difference is which server list they hand to register_server_tests()):
os-servers.yml's job selects the native-server subset by filename
(unittest discover -p), so the file boundary is a CI-selection
mechanism, not a content difference.

Touches CI: os-servers.yml and .github/actions/os-server/* reference
the native file by name; servers.mk's test-servers/test-os-server make
targets reference both filenames via -p. Verified: flake8 clean, every
functional module imports, and the unit suite passes unchanged (270
tests, 2 expected failures).
