+++ tests/functional/test_roundtrip.py
@@ -9,7 +9,7 @@ from unittest import TestCase
 from PIL import Image
 
 from .test_proxy import _await_capture, _start_vnclog, _stop_proxy
-from .vncservers import HOST, TIGERVNC_AUTH, VNCDO_REPLAY, port_open, start_replay_server, vncdo_argv
+from .utils import HOST, TIGERVNC_AUTH, VNCDO_REPLAY, port_open, start_replay_server, vncdo_argv
 
 ROUNDTRIP_PROXY_PORT = 5996
 REPLAY_PORT = 5997
